feat: expose customDocs prop on API
