+++ packages/elements/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@abbudao/elements",
-  "version": "7.8.0",
+  "version": "7.9.0",
   "description": "UI components for composing beautiful developer documentation.",
   "keywords": [],
   "sideEffects": [
