+++ packages/elements/src/components/API/APIWithStackedLayout.tsx
@@ -12,8 +12,9 @@ import cn from 'classnames';
 import * as React from 'react';
 import { useLocation } from 'react-router-dom';
 
-import { OperationNode, ServiceNode } from '../../utils/oas/types';
-import { computeTagGroups, TagGroup } from './utils';
+import type { OperationNode, ServiceNode } from '../../utils/oas/types';
+import type { TagGroup } from './types';
+import { computeTagGroups } from './utils';
 
 type TryItCredentialsPolicy = 'omit' | 'include' | 'same-origin';
